// src/types/catalogFirst.ts
export interface ICatalogFirst {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: any;
  is_hidden_in_menu: any;
  path: string;
  path_for_pli: string;
  level: number;
  type: any;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: any;
  level2: any;
  level3: any;
  level4: any;
  level5: any;
  level6: any;
  level7: any;
  level8: any;
  level9: any;
  level10: any;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  id_cat: number;
  full_slug2: string;
}