// src/types/catalogSecond.ts
export interface ICatalogSecond {
  full_slug: string;
  model: Model[];
  category: Category;
  totalCount: number;
  model_count: number;
}

export interface Model {
  id: number;
  art3: any;
  art: number;
  cena0: string;
  cena4: any;
  cena_dos?: string;
  cenaok: number;
  cena0r: any;
  cena4r: any;
  cena_dosr: any;
  cena_kyrs: any;
  naim: string;
  url: string;
  prim?: string;
  img?: string;
  idt?: string;
  notfound: number;
  id_city: number;
  dat1: any;
  minQty: number;
  isNovelty?: number;
  country_id: any;
  country?: string;
  stuff?: string;
  size: string;
  keep_package: any;
  is_paid_delivery: any;
  discountPercent: any;
  currencySign: string;
  supply_period: any;
  balance: string;
  id_post: number;
  id_cat: number;
  bazedin: string;
  id_cat1c: any;
  naim_cat1c: any;
  id_idcat: any;
  moder: number;
  id_tov: number;
  copy: any;
  weight: any;
  description: string;
  short_description: string;
  trademark: string;
  cert?: string;
  pli?: number;
  naim_word?: string;
  img1sm: string;
  img2big: string;
  art_post: string;
  id_user_add?: number;
  price_cost?: string;
  price_update?: string;
  naim_add?: string;
  naim_add_manual?: string;
  id_user_add_manual?: number;
  id_status: string;
  video?: string;
  pli_update?: string;
  meta_title?: string;
  meta_description?: string;
  meta_keywords?: string;
  trademark_id: any;
  in_box: any;
  box: any;
  img_url: any;
  status: number;
  id_micro_serv: any;
  d_min: any;
  d_max: any;
  percent_sum: string;
  update_period: number;
  show_price: number;
  show_opt: number;
  active_img?: number;
  active_percent_sum: number;
  images: Image[];
  price: number;
  before_markup: number;
  before_round?: number;
  round_test?: string;
  apply_test_per: string;
  discount: number;
  discount_prc: number;
  promotions: any[];
  old_price: number;
  to_date: number;
  from_date: number;
  valuteVal: string;
  date_dost: string;
  photos: Photo[];
}

export interface Image {
  id: number;
  id_tov: number;
  img: string;
  sort: number;
  id_post: number;
  id_tovv: number;
}

export interface Photo {
  url_part: string;
}

export interface Category {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: any;
  is_hidden_in_menu: any;
  path: string;
  path_for_pli: string;
  level: number;
  type: any;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: any;
  level2: any;
  level3: any;
  level4: any;
  level5: any;
  level6: any;
  level7: any;
  level8: any;
  level9: any;
  level10: any;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: Tov[];
  "0": N0;
  "1": N1;
  "2": N2;
  "3": N3;
  "4": N4;
  "5": N5;
  "6": N6;
  "7": N7;
  "8": N8;
  "9": N9;
  "10": N10;
  "11": N11;
  "12": N12;
  "13": N13;
}

export interface Tov {
  id: number;
  art3: any;
  art: number;
  cena0: string;
  cena4: any;
  cena_dos?: string;
  cenaok: number;
  cena0r: any;
  cena4r: any;
  cena_dosr: any;
  cena_kyrs: any;
  naim: string;
  url: string;
  prim?: string;
  img?: string;
  idt?: string;
  notfound: number;
  id_city: number;
  dat1: any;
  minQty: number;
  isNovelty?: number;
  country_id: any;
  country?: string;
  stuff?: string;
  size: string;
  keep_package: any;
  is_paid_delivery: any;
  discountPercent: any;
  currencySign: string;
  supply_period: any;
  balance: string;
  id_post: number;
  id_cat: number;
  bazedin: string;
  id_cat1c: any;
  naim_cat1c: any;
  id_idcat: any;
  moder: number;
  id_tov: number;
  copy: any;
  weight: any;
  description: string;
  short_description: string;
  trademark: string;
  cert?: string;
  pli?: number;
  naim_word?: string;
  img1sm: string;
  img2big: string;
  art_post: string;
  id_user_add?: number;
  price_cost?: string;
  price_update?: string;
  naim_add?: string;
  naim_add_manual?: string;
  id_user_add_manual?: number;
  id_status: string;
  video?: string;
  pli_update?: string;
  meta_title?: string;
  meta_description?: string;
  meta_keywords?: string;
  trademark_id: any;
  in_box: any;
  box: any;
  img_url: any;
  status: number;
  id_micro_serv: any;
  d_min: any;
  d_max: any;
  percent_sum: string;
  update_period: number;
  show_price: number;
  show_opt: number;
  active_img?: number;
  active_percent_sum: number;
  images: Image2[];
  price: number;
  before_markup: number;
  before_round?: number;
  round_test?: string;
  apply_test_per: string;
  discount: number;
  discount_prc: number;
  promotions: any[];
  old_price: number;
  to_date: number;
  from_date: number;
  valuteVal: string;
  date_dost: string;
  photos: Photo2[];
}

export interface Image2 {
  id: number;
  id_tov: number;
  img: string;
  sort: number;
  id_post: number;
  id_tovv: number;
}

export interface Photo2 {
  url_part: string;
}

export interface N0 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: string;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: Tov2[];
}
// ...
export interface N1 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: string;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: string;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: any[];
}

export interface N2 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: string;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: string;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: any[];
}

export interface N3 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: Tov3[];
}
// ...
export interface N4 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: any;
  seo_keyword: any;
  parent_for_pli: any;
  tov: Tov4[];
}
// ...
export interface N5 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: Tov5[];
}
// ...
export interface N6 {
  idd: number;
  id: number;
  sid: number;
  name: string;
  priority: number;
  priority_home: number;
  priority_menu: number;
  is_hidden_in_menu: number;
  path: string;
  path_for_pli: string;
  level: number;
  type: number;
  photo: any;
  icon: string;
  top_image: any;
  is_leaf: number;
  full_slug_old: string;
  full_slug: string;
  level1: number;
  level2: number;
  level3: number;
  level4: number;
  level5: number;
  level6: number;
  level7: number;
  level8: number;
  level9: number;
  level10: number;
  nal: number;
  opis: string;
  html: any;
  nazv: string;
  title: string;
  description: string;
  keywords: string;
  count_skid: number;
  count_hit: number;
  parent: number;
  e_id: any;
  main_type: any;
  sort: number;
  sort_menu: number;
  manual: number;
  enable: number;
  sim_enable: number;
  status: any;
  seo_title: string;
  seo_description: string;
  seo_keyword: string;
  parent_for_pli: any;
  tov: Tov6[];
}